+++ primal/src/Control/Prim/Monad.hs
@@ -9,8 +9,14 @@
 module Control.Prim.Monad
   ( module Control.Prim.Monad.Internal
   , touch
-  , evaluate
   , keepAlive
+  , eval
+  , evalM
+  , NFData
+  , deepeval
+  , deepevalM
+  , whenM
+  , unlessM
   -- * Re-export
   , module Control.Monad
   ) where
@@ -19,3 +25,20 @@ import GHC.Exts
 import Control.Prim.Eval
 import Control.Prim.Monad.Internal
 import Control.Monad
+
+
+
+-- | Similar to `when`, but condional is supplied in a form of monadic action rather than a
+-- pure value.
+--
+-- @since 0.3.0
+whenM :: Monad m => m Bool -> m () -> m ()
+whenM m action = m >>= \b -> when b action
+
+
+-- | Similar to `unless`, but condional is supplied in a form of monadic action rather than a
+-- pure value.
+--
+-- @since 0.3.0
+unlessM :: Monad m => m Bool -> m () -> m ()
+unlessM m action = m >>= \b -> unless b action