+++ primal/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.3.0
 
+* Addition of `eval`, `evalM`, `deepeval` and `deepevalM`
+* Addittion of `whenM` and `unlessM`
 * Whole bunch of concurrency and exception functionality
 * Addition of `Ref` adnd `MVar`
 * Addition of basic array functionality: